Revert "观察者模式-基于Sping状态机-end"
This reverts commit f57502308ff54a95f4306878960f962b94f98def.

+++ src/main/java/com/coderwhs/designPattern/model/entity/Order.java
@@ -3,8 +3,6 @@ package com.coderwhs.designPattern.model.entity;
 import com.coderwhs.designPattern.model.enums.OrderStateEnum;
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 @NoArgsConstructor
@@ -14,6 +12,6 @@ public class Order {
     private String orderId;
     private String productId;
     private OrderStateEnum orderState;//订单状态
-    private BigDecimal price;//商品价格
+    private Float price;//商品价格
     private String userId;//当前用户唯一Id
 }
